refactor dpm campaign definition routes
